Multiple Choices
added the option to let the user pick multiple dice rolls at once. This was from a dropdown box mixed with a button. Needed to update the bootstrap library to use that function. The log area is now renamed results and it got a visual overhaul. Will work on code reusability next update.